moved instructions into client mod
moved packet_id and instruction_id in bus mod
