+++ src/client/instructions/action.rs
@@ -1,5 +1,5 @@
 use super::Client;
-use crate::{instruction_id, packet_id};
+use crate::bus::{instruction_id, packet_id};
 use crate::{Response, TransferError, WriteError};
 
 #[super::bisync]
@@ -16,7 +16,7 @@ where
 	/// If you want to broadcast this instruction, it may be more convenient to use [`Self::broadcast_action()`] instead.
 	pub async fn action(&mut self, motor_id: u8) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::ACTION, 0, |_| Ok(())).await?;
-		Ok(self.read_response_if_not_broadcast( motor_id).await?)
+		Ok(self.read_response_if_not_broadcast(motor_id).await?)
 	}
 
 	/// Broadcast an action command to all connected motors to trigger a previously registered instruction.
