+++ src/client/instructions/bulk_write.rs
@@ -1,7 +1,8 @@
 use super::Client;
 use crate::bus::endian::{write_u16_le, write_u8_le};
+use crate::bus::{instruction_id, packet_id};
+use crate::BulkWriteData;
 use crate::WriteError;
-use crate::{instruction_id, packet_id, BulkWriteData};
 
 #[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
