+++ src/client/instructions/clear.rs
@@ -1,5 +1,5 @@
 use super::Client;
-use crate::{instruction_id, packet_id};
+use crate::bus::{instruction_id, packet_id};
 use crate::{Response, TransferError, WriteError};
 
 /// The parameters for the CLEAR command to clear the revolution counter.
@@ -62,7 +62,7 @@ where
 	pub async fn clear_error(&mut self, motor_id: u8) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::CLEAR, CLEAR_ERROR.len(), clear_error_parameters)
 			.await?;
-		Ok(self.read_response_if_not_broadcast( motor_id).await?)
+		Ok(self.read_response_if_not_broadcast(motor_id).await?)
 	}
 
 	/// Try to clear the error of all motors on the bus.
