+++ src/client/instructions/reg_write.rs
@@ -1,7 +1,7 @@
 use super::Client;
 use crate::bus::endian::write_u16_le;
+use crate::bus::instruction_id;
 use crate::bus::Data;
-use crate::instruction_id;
 use crate::{Response, TransferError};
 
 #[super::bisync]
@@ -31,7 +31,7 @@ where
 			Ok(())
 		})
 		.await?;
-		Ok(self.read_response_if_not_broadcast( motor_id).await?)
+		Ok(self.read_response_if_not_broadcast(motor_id).await?)
 	}
 
 	/// Register a write command for value to a specific motor.
@@ -54,6 +54,6 @@ where
 			value.encode(&mut buffer[2..])
 		})
 		.await?;
-		Ok(self.read_response_if_not_broadcast( motor_id).await?)
+		Ok(self.read_response_if_not_broadcast(motor_id).await?)
 	}
 }
