+++ src/common/mod.rs
@@ -1,132 +1,5 @@
-use crate::{bus::StatusPacket, Response};
-
-/// Raw instructions IDs.
-#[rustfmt::skip]
-#[allow(missing_docs)]
-pub mod instruction_id {
-	pub const PING          : u8 = 0x01;
-	pub const READ          : u8 = 0x02;
-	pub const WRITE         : u8 = 0x03;
-	pub const REG_WRITE     : u8 = 0x04;
-	pub const ACTION        : u8 = 0x05;
-	pub const FACTORY_RESET : u8 = 0x06;
-	pub const REBOOT        : u8 = 0x08;
-	pub const CLEAR         : u8 = 0x10;
-	pub const SYNC_READ     : u8 = 0x82;
-	pub const SYNC_WRITE    : u8 = 0x83;
-	pub const BULK_READ     : u8 = 0x92;
-	pub const BULK_WRITE    : u8 = 0x93;
-	pub const STATUS        : u8 = 0x55;
-}
-
-/// Special packet IDs.
-pub mod packet_id {
-	/// The broadcast address.
-	pub const BROADCAST: u8 = 0xFE;
-}
-
-/// Sync data for a specific motor.
-///
-/// Used by [`crate::Client::sync_write`] and [`crate::Client::sync_write_bytes`]
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct SyncWriteData<T> {
-	/// The ID of the motor.
-	pub motor_id: u8,
-
-	/// The data to be written to the motor.
-	pub data: T,
-}
-
-impl<T> AsRef<SyncWriteData<T>> for SyncWriteData<T> {
-	fn as_ref(&self) -> &Self {
-		self
-	}
-}
-
-/// Bulk data for a specific motor.
-///
-/// This struct is similar to [`SyncWriteData`],
-/// but it supports reads and writes of different sizes and to different addresses for each motor.
-///
-/// Used with [`crate::Client::bulk_write`].
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct BulkWriteData<T> {
-	/// The ID of the motor.
-	pub motor_id: u8,
-
-	/// The address to be written to
-	pub address: u16,
-
-	/// The data to be written
-	pub data: T,
-}
-
-impl<T> AsRef<BulkWriteData<T>> for BulkWriteData<T> {
-	fn as_ref(&self) -> &Self {
-		self
-	}
-}
-
-/// Parameters for a bulk read instruction.
-///
-/// Use with [`crate::Client::bulk_read_bytes`].
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub struct BulkReadData {
-	/// The ID of the motor.
-	pub motor_id: u8,
-
-	/// The address of the data to be read
-	pub address: u16,
-
-	/// The length of the data to be read.
-	pub count: u16,
-}
-
-/// The kind of factory reset to perform.
-#[repr(u8)]
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub enum FactoryResetKind {
-	/// Reset all settings, including the motor ID and baud rate.
-	ResetAll = 0xFF,
-
-	/// Reset all settings except for the motor ID.
-	KeepId = 0x01,
-
-	/// Reset all settings except for the motor ID and baud rate.
-	KeepIdAndBaudRate = 0x02,
-}
-
-/// A response from a motor to a ping instruction.
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Ping {
-	/// The model of the motor.
-	///
-	/// Refer to the online manual to find the codes for each model.
-	pub model: u16,
-
-	/// The firmware version of the motor.
-	pub firmware: u8,
-}
-
-impl<'a> TryFrom<StatusPacket<'a>> for Response<Ping> {
-	type Error = crate::InvalidParameterCount;
-
-	fn try_from(status_packet: StatusPacket<'a>) -> Result<Self, Self::Error> {
-		let parameters = status_packet.parameters();
-		crate::InvalidParameterCount::check(parameters.len(), 3)?;
-		Ok(Self {
-			motor_id: status_packet.packet_id(),
-			alert: status_packet.alert(),
-			data: Ping {
-				model: crate::bus::endian::read_u16_le(&parameters[0..]),
-				firmware: crate::bus::endian::read_u8_le(&parameters[2..]),
-			},
-		})
-	}
-}
-
 use crate::bus::endian::read_u16_le;
-use crate::bus::InstructionPacket;
+use crate::bus::{instruction_id, InstructionPacket};
 use crate::InvalidParameterCount;
 
 #[cfg(feature = "alloc")]
